requirements.txt
beautifulsoup4==4.5.1
click==6.6
colorama==0.3.7
cookies==2.2.1
cycler==0.10.0
decorator==4.0.10
entrypoints==0.2.2
flask>=0.12
geopy==1.11.0
googlefinance==0.7
httplib2==0.9.2
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.1.0
notebook==4.2.3
numpy==1.11.2
oauth2==1.9.0.post1
pandas==0.19.0
pickleshare==0.7.4
prompt-toolkit==1.0.8
Pygments==2.1.3
pyparsing==2.1.10
python-dateutil==2.5.3
python-dotenv==0.6.1
python-forecastio==1.3.5
pytz==2016.7
pyzmq==16.0.0
qtconsole==4.2.1
requests==2.12.1
responses==0.5.1
scikit-learn==0.18
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
tornado==4.4.2
traitlets==4.3.1
twilio~=6.0.0
wcwidth==0.1.7
Werkzeug==0.11.11
widgetsnbextension==1.2.6
win-unicode-console==0.5
yelp==1.0.2
